:wrench: Creates game and hud classes
